fix(pm): 性能仪表盘对含 NaN 聚合结果返回空 body(200) 致全白
根因：PM 结果行 metric_value(float64) 含 NaN/Inf 时 Go encoding/json 编码
报错；Gin 先写 200 头再流式编码 body，编到 NaN 行中断 → 客户端收到 200+空
body → 前端「暂无数据」。平均/比率型指标分母为 0 极易算出 NaN，只要某任务含
一个 NaN，仪表盘该任务即全白。

修复：在共享响应封装层（OK/OKWithStatus/OKWithMsg）统一加 sanitize pass，
c.JSON 前把 payload 内所有非有限浮点（NaN/±Inf）规整为 JSON null（缺测语义，
不伪造成 0）。含保真守门：不含非有限浮点的响应原样返回入参、零拷贝零副作用，
仅真含 NaN 的罕见 PM 数值行走净化。所有走信封的 PM 端点（adhoc results、
metrics/aggregated、counters、counters/aggregated、kpis）一并受益，新端点免疫。

回归测试覆盖两条路径：全有限值正常返回 + 含 NaN/Inf 返回 200+该字段 null。
运行栈实测：含 NaN 的内置「全网-LTE」任务 results(limit=10000) 由 0 字节恢复
为 1.75MB，2 行 NaN 精确呈现 null 无裸 NaN 漏网，仪表盘出图正常。

Closes #387

+++ omcgo/internal/pm/adhoc/executor.go
@@ -192,7 +192,7 @@ func (e *Executor) queryAndConvert(ctx context.Context, task *Task, g metrics.Gr
 			ProductID:   r.ProductID, // T-0182-fix: product 维度分组键透传（device/aggregate_group 维度为 Nil）
 			MetricPath:  r.MetricPath,
 			MetricType:  string(r.MetricType),
-			MetricValue: r.MetricValue,
+			MetricValue: float64(r.MetricValue),
 			StatisType:  statisStr,
 			Granularity: string(r.Granularity),
 			Time:        r.Time,
